post routes for place and attraction, disabled addplace button

+++ migrations/20190116180718-create-place.js
@@ -12,10 +12,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       image: {
         type: Sequelize.TEXT
